+++ exercise-microservice/src/main/java/pl/edu/pg/eti/exercisemicroservice/ExerciseMicroserviceApplication.java
@@ -1,0 +1,13 @@
+package pl.edu.pg.eti.exercisemicroservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ExerciseMicroserviceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ExerciseMicroserviceApplication.class, args);
+    }
+
+}
