+++ exercise-microservice/src/main/java/pl/edu/pg/eti/exercisemicroservice/exercises/dto/PutPatchExerciseRequest.java
