+++ exercise-microservice/src/main/java/pl/edu/pg/eti/exercisemicroservice/exercises/entity/Exercise.java
