+++ exercise-microservice/src/main/java/pl/edu/pg/eti/exercisemicroservice/exercises/repository/api/ExerciseRepository.java
