+++ exercise-microservice/src/test/java/pl/edu/pg/eti/exercisemicroservice/ExerciseMicroserviceApplicationTests.java
@@ -1,0 +1,13 @@
+package pl.edu.pg.eti.exercisemicroservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ExerciseMicroserviceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
