--- requests.http
@@ -1,53 +1,0 @@
-### Get all workouts
-GET http://localhost:8080/api/workouts
-
-### Get particular workout
-GET http://localhost:8080/api/workouts/de1eaf7b-1b68-4c0a-a15e-682da5aa4c66
-
-### Delete workout
-DELETE http://localhost:8080/api/workouts/de1eaf7b-1b68-4c0a-a15e-682da5aa4c66
-
-### Put workout
-PUT http://localhost:8080/api/workouts/4c99a53b-5590-4e9c-9b22-e886e89e48cd
-Content-Type: application/json
-
-{
-  "name": "Arm day",
-  "duration": 100,
-  "estimated_calories": 700
-}
-
-### Patch workout
-PATCH http://localhost:8080/api/workouts/4c99a53b-5590-4e9c-9b22-e886e89e48cd
-Content-Type: application/json
-
-{
-  "name": "Biceps day"
-}
-
-
-### Get Workout's exercises
-GET http://localhost:8080/api/workouts/de1eaf7b-1b68-4c0a-a15e-682da5aa4c66/exercises
-
-### Delete exercise from workout
-DELETE http://localhost:8080/api/workouts/de1eaf7b-1b68-4c0a-a15e-682da5aa4c66/exercises/00129945-fdb5-4027-b86b-ad1be26f6376
-
-
-### Put exercise to workout
-PUT http://localhost:8080/api/workouts/de1eaf7b-1b68-4c0a-a15e-682da5aa4c66/exercises/051f2d79-e7e0-4364-a5a0-41551884ce48
-Content-Type: application/json
-
-{
-  "name": "Back Squats",
-  "reps": 10,
-  "sets": 4,
-  "additional_weight": 40
-}
-
-### Patch exercise to workout
-PATCH http://localhost:8080/api/workouts/de1eaf7b-1b68-4c0a-a15e-682da5aa4c66/exercises/051f2d79-e7e0-4364-a5a0-41551884ce48
-Content-Type: application/json
-
-{
-  "additional_weight": 90
-}
