--- src/main/java/pl/edu/pg/eti/labproject/SimpleWorkoutApplication.java
@@ -1,13 +1,0 @@
-package pl.edu.pg.eti.labproject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SimpleWorkoutApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SimpleWorkoutApplication.class, args);
-    }
-
-}
