--- src/main/java/pl/edu/pg/eti/labproject/function/TriFunction.java
@@ -1,16 +1,0 @@
-package pl.edu.pg.eti.labproject.function;
-
-import java.util.Objects;
-import java.util.function.Function;
-
-@FunctionalInterface
-public interface TriFunction<A,B,C,R> {
-
-    R apply(A a, B b, C c);
-
-    default <V> TriFunction<A, B, C, V> andThen(
-            Function<? super R, ? extends V> after) {
-        Objects.requireNonNull(after);
-        return (A a, B b, C c) -> after.apply(apply(a, b, c));
-    }
-}
