--- src/main/resources/application.properties
@@ -1,17 +1,0 @@
-server.port=7890
-spring.application.name=simple-workout
-
-spring.datasource.url=jdbc:h2:mem:simple-workout
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=admin
-spring.datasource.password=password
-
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-spring.jpa.generate-ddl=true
-
-# TODO: change after db migrations will be added
-spring.jpa.hibernate.ddl-auto=update
-
-spring.h2.console.enabled=true
-
-spring.jackson.property-naming-strategy=SNAKE_CASE
