--- src/test/java/pl/edu/pg/eti/labproject/LabProjectApplicationTests.java
@@ -1,13 +1,0 @@
-package pl.edu.pg.eti.labproject;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LabProjectApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
