+++ workout-microservice/src/main/java/pl/edu/pg/eti/workoutmicroservice/WorkoutMicroserviceApplication.java
@@ -1,0 +1,13 @@
+package pl.edu.pg.eti.workoutmicroservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WorkoutMicroserviceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(WorkoutMicroserviceApplication.class, args);
+    }
+
+}
