+++ workout-microservice/src/test/java/pl/edu/pg/eti/workoutmicroservice/WorkoutMicroserviceApplicationTests.java
@@ -1,0 +1,13 @@
+package pl.edu.pg.eti.workoutmicroservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class WorkoutMicroserviceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
